adding object oriented interfaces and seperating filtration building from diagram solving
