+++ cechmate/__init__.py
@@ -1,2 +1,3 @@
 from .visuals import *
 from .filtrations import *
+from .solver import *
