+++ test/test_filtrations.py
@@ -2,7 +2,20 @@ import time
 
 import numpy as np
 
-from cechmate import plot_diagram, alpha_filtration, rips_filtration
+from cechmate import plot_diagram, phat_diagrams, Alpha, Rips #alpha_filtration, rips_filtration
+
+
+def test_phat_diagrams():
+
+    t = np.linspace(0, 2 * np.pi, 40)
+    X = np.zeros((len(t), 2))
+    X[:, 0] = np.cos(t)
+    X[:, 1] = np.sin(t)
+    np.random.seed(10)
+    X += 0.2 * np.random.randn(len(t), 2)
+    rips = Rips(1).build(X)
+
+    dgms = phat_diagrams(rips)
 
 
 def test_rips():
@@ -15,7 +28,7 @@ def test_rips():
     X[:, 1] = np.sin(t)
     np.random.seed(10)
     X += 0.2 * np.random.randn(len(t), 2)
-    rips = rips_filtration(X, 1)
+    rips = Rips(1).build(X)
 
 
 def test_alpha():
@@ -24,5 +37,5 @@ def test_alpha():
     X = np.random.randn(50, 4)
     X = X / np.sqrt(np.sum(X ** 2, 1)[:, None])
     tic = time.time()
-    diagrams = alpha_filtration(X)
+    diagrams = Alpha().build(X)
     phattime = time.time() - tic
